requirements.txt
Flask
psycopg2-binary
bcrypt
PyJWT
python-dotenv
gunicorn
flask-cors
